remove index route to prevent unnecessary redirect
