+++ apps/web/.env
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_DEFAULT_ROOT_COMPONENT=andyh.near/widget/LandingPage
